Finalizando reparos

+++ src/main/java/br/gov/rn/saogoncalo/smtsis/services/BenfeitoriasService.java
@@ -14,24 +14,36 @@ public class BenfeitoriasService {
 	@Autowired
 	private BenfeitoriasRepository repository;
 	
-	public void Salvar(Benfeitorias benfeitoria) {
-		repository.save(benfeitoria);
-	}
+	public Benfeitorias salvar(Benfeitorias benfeitorias) {
+         return repository.save(benfeitorias);
+    }
 
 	public List<Benfeitorias> BuscarTodos() {
 		return repository.findAll();
 	}
 
-	public Benfeitorias Buscar(Long id) {
+	public Benfeitorias BuscarPorId(Long id) {
 		return repository.findById(id).get();
 	}
-
-	public void remover(Benfeitorias benfeitoria) {
-		repository.delete(benfeitoria);
-	}
-
-	public void removerPorId(Long id) {
-		repository.deleteById(id);
-	}
 	
+	public Boolean deleta(Long id){
+        if (benfeitoriaExiste(id)){
+        	repository.deleteById(id);
+            return true;
+        }
+        return false;
+    }
+
+    private boolean benfeitoriaExiste(Long id){
+        return repository.existsById(id);
+    }
+	
+    public Benfeitorias editar(Benfeitorias benfeitoria) {
+        if (benfeitoria.getId() != null){
+            if (benfeitoriaExiste(benfeitoria.getId())){
+                return repository.save(benfeitoria);
+            }
+        }
+        return null;
+    }
 }
